Microbe pushing kinda works
Omg it hurts how much time it wasted

+++ Assets/Scripts/Core/Ingame/IngameLevelHandler.cs
@@ -32,7 +32,7 @@ namespace Core
             }
         }
 
-        public float GetCurrentLevel()
+        public int GetCurrentLevel()
         {
             return level;
         }
